pyproject.toml
[tool.poetry]
name = "blockchain"
version = "0.1.0"
description = ""
authors = ["gaulgeous <uqdgaul@uq.edu.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
